// enum/statusCode/statusCode.enum.ts
export enum STATUSCODE {
    // Sucesso
    OK = 200,
    CREATED = 201,
    NO_CONTENT = 204,

    // Redirecionamento
    MOVED_PERMANENTLY = 301,
    FOUND = 302,

    // Erros do Cliente
    BAD_REQUEST = 400,
    UNAUTHORIZED = 401,
    FORBIDDEN = 403,
    NOT_FOUND = 404,
    CONFLICT = 409,

    // Erros do Servidor
    INTERNAL_SERVER_ERROR = 500,
    NOT_IMPLEMENTED = 501,
    BAD_GATEWAY = 502,
    SERVICE_UNAVAILABLE = 503,
}